fix - long

+++ libretaDigital/src/main/java/com/libretaDigital/beans/AbsenceBean.java
@@ -7,13 +7,13 @@ import org.codehaus.jackson.annotate.JsonAutoDetect;
 @JsonAutoDetect
 public class AbsenceBean {
 
-	private BigInteger idStudent;	
+	private long idStudent;	
 	private boolean late;;	
 
-	public BigInteger getIdStudent() {
+	public long getIdStudent() {
 		return idStudent;
 	}
-	public void setIdStudent(BigInteger idStudent) {
+	public void setIdStudent(long idStudent) {
 		this.idStudent = idStudent;
 	}
 	public boolean isLate() {
